@@ -1,44 +1,41 @@
 {
-  "name": "edu-replay",
-  "version": "1.0.0",
-  "description": "",
-  "main": "main.js",
-  "scripts": {
-    "start": "npm run build:dev && electron .",
-    "build:dev": "cross-env node_env=test webpack --config webpack.config.js",
-    "watch:dev": "cross-env node_env=test webpack --watch --config webpack.config.js",
-    "build:win": "npm run build:dev && cross-env node_env=production electron-packager ./ whiteboard-electron --platform=win32 --arch=x64 --overwrite",
-    "build:mac": "npm run build:dev && cross-env node_env=production electron-packager ./ whiteboard-electron --platform=darwin --arch=x64 --overwrite"
-  },
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "@babel/core": "^7.13.16",
-    "@babel/preset-env": "^7.13.15",
-    "@babel/preset-react": "^7.13.13",
-    "@types/react": "^17.0.11",
-    "autoprefixer": "^10.2.5",
-    "awesome-typescript-loader": "^5.2.1",
-    "babel-loader": "^8.2.2",
-    "babel-polyfill": "^6.26.0",
-    "cross-env": "^7.0.3",
-    "css-loader": "^5.2.4",
-    "electron": "^13.1.4",
-    "electron-packager": "^15.2.0",
-    "electron-reloader": "^1.2.1",
-    "html-webpack-plugin": "^5.3.1",
-    "postcss-loader": "^5.2.0",
-    "precss": "^4.0.0",
-    "react-router-dom": "^5.2.0",
-    "style-loader": "^2.0.0",
-    "typescript": "^4.2.4",
-    "webpack": "^5.35.1",
-    "webpack-cli": "^4.6.0",
-    "webpack-dev-server": "^3.11.2"
-  },
-  "dependencies": {
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "sha1": "^1.1.1"
-  }
+	"name": "whiteboard",
+	"version": "1.0.0",
+	"description": "whiteboard demo",
+	"main": "main.js",
+	"scripts": {
+		"start": "npm run build:dev && electron .",
+		"build:dev": "cross-env node_env=test webpack --config webpack.config.js",
+		"watch:dev": "cross-env node_env=test webpack --watch --config webpack.config.js",
+		"build:win": "npm run build:dev && cross-env node_env=production electron-packager ./ whiteboard-electron --platform=win32 --arch=x64 --overwrite",
+		"build:mac": "npm run build:dev && cross-env node_env=production electron-packager ./ whiteboard-electron --platform=darwin --arch=x64 --overwrite"
+	},
+	"license": "ISC",
+	"dependencies": {
+		"react": "^18.2.0",
+		"react-dom": "^18.2.0",
+		"react-router-dom": "^6.4.3"
+	},
+	"devDependencies": {
+		"@babel/core": "^7.19.6",
+		"@babel/preset-env": "^7.19.4",
+		"@types/react": "^18.0.24",
+		"@types/react-dom": "^18.0.8",
+		"babel-loader": "^9.1.0",
+		"cross-env": "^7.0.3",
+		"electron": "^13.1.4",
+		"electron-packager": "^15.2.0",
+		"electron-reloader": "^1.2.1",
+		"css-loader": "^6.7.1",
+		"html-webpack-plugin": "^5.5.0",
+		"less": "^4.1.3",
+		"less-loader": "^11.1.0",
+		"mini-css-extract-plugin": "^2.6.1",
+		"ts-loader": "^9.4.1",
+		"typescript": "^4.8.4",
+		"webpack": "^5.74.0",
+		"webpack-bundle-analyzer": "^4.7.0",
+		"webpack-cli": "^4.10.0",
+		"webpack-dev-server": "^4.11.1"
+	}
 }
