@@ -17,7 +17,7 @@ function createWindow () {
   })
   mainWindow.maximize()
   // and load the index.html of the app.
-  mainWindow.loadFile('index.html')
+  mainWindow.loadFile('./dist/assets/index.html')
 
   // Open the DevTools.
   // mainWindow.webContents.openDevTools()
